+++ database/seeders/DatabaseSeeder.php
@@ -2,9 +2,11 @@
 
 namespace Database\Seeders;
 
+use App\Enum\Users\RoleEnum;
 use App\Models\User;
-// use Illuminate\Database\Console\Seeds\WithoutModelEvents;
 use Illuminate\Database\Seeder;
+
+// use Illuminate\Database\Console\Seeds\WithoutModelEvents;
 
 class DatabaseSeeder extends Seeder
 {
@@ -18,6 +20,10 @@ class DatabaseSeeder extends Seeder
         User::factory()->create([
             'name' => 'Test User',
             'email' => 'test@example.com',
+        ])->assignRole(RoleEnum::SUPER_ADMIN->value);
+
+        $this->call([
+            ShieldSeeder::class,
         ]);
     }
 }